# backend/services/notification-service/src/main/java/com/smartmarket/notification/application/dto/GeofenceEvaluateResponse.java
package com.smartmarket.notification.application.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class GeofenceEvaluateResponse {
    private boolean accepted;
    private String traceId;
}