# backend/services/notification-service/src/main/java/com/smartmarket/notification/application/dto/PageMetadata.java
package com.smartmarket.notification.application.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class PageMetadata {
    private int page;
    private int size;
    private long totalElements;
    private int totalPages;
}